Call device.connect() aftre creating NukiDevice
In latest pyNukiBT version, we need to call device.connect() so the
correct nuki device_type can be identified. This was done to improve the
retry mechanism.

Plus, this way we can make sure HA's BT is able to connect to the Nuki
device before we try more complex operations. And can fail properly if
no connection can be made.